requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
cryptography>=41.0.0
